Renamed AllocationsMenu and ReceiptsMenu to just Allocations and Receipts. Moved related files into ReceiptManagement and AllocationManagement folders. Separated upload functionality out of DocumentsList and into DocumentUpload
